Rename pickups to use original tech name
